requirements.txt
coreapi
django
django-cors-headers
django-environ
django-heroku
djangorestframework
gunicorn
pre-commit
